axi_dmac: Remove length alignment requirements for MM interfaces
Currently the DMAC requires that the length of the transfer is aligned to
the widest interface. This is done since it is not possible to do partial
beats or fractional clock cycles.

For memory mapped interfaces it is possible to workaround this restriction
though.

For reading from system memory a full beat is transferred, but the MSBs of
the beat are discarded.

For writing to system memory a full beat is transferred, but the wstrb
signal is used to indicate that the MSBs of the beat should not be
committed to system memory.

This will affect the throughput since the MSBs of the beat do not carry any
meaningful data. This effect should be small though since this only happens
for the last beat of a transfer.

The length alignment requirement for other interfaces still exists though.
E.g. for a DMAC with a 64-bit memory interface and a 16-bit FIFO interface
the length alignment requirement will be 2 bytes. Whereas for a DMAC with a
64-bit memory interface and a 128-bit FIFO interface the length alignment
requirement will be 8 bytes.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>