Add project for the AD9136-FMC-EBZ + ZC706
Add a project for the AD9136-FMC-EBZ dual DAC board connected to the ZC706
carrier board.

This project assumes that the transceiver reference clock and SYSREF are
provided via the clock distribution chip that is found on the
AD9136-FMC-EBZ board. The reference clock rate must be the lane rate
divided by 40. SYSREF should be the reference clock divided by 64 (or a
larger power-of-two).

Due to system constraints the reference clock must be in the range of 75 to
165 MHz, which means the lanerate must be between 3 and 6.6 Gbps

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>